Fine-tuning Project/rag-lite/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
transformers==4.39.3
peft==0.10.0
datasets==2.18.0
sentence-transformers==2.7.0
torch>=2.0.0
accelerate>=0.21.0
bitsandbytes>=0.41.0

# Vector database
faiss-cpu==1.7.2

# Text processing
nltk>=3.8.0
beautifulsoup4>=4.12.0

# API and utilities
fastapi==0.110
uvicorn==0.29
python-dotenv>=1.0.0

# Training and optimization
tensorboard>=2.14.0

